Moves data drops out into a new module (dlg.data)
Moves BranchAppDrop into dlg.apps.

Moves json_drop into this directory as well.
Updates all related tests and usages.
