+++ daliuge-engine/dlg/data/memory.py
@@ -1,0 +1,120 @@
+import base64
+import io
+import os
+import random
+import string
+import sys
+
+from dlg.common.reproducibility.reproducibility import common_hash
+from dlg.drop import DataDROP, logger
+from dlg.io import SharedMemoryIO, MemoryIO
+
+
+##
+# @brief Memory
+# @details In-memory storage of intermediate data products
+# @par EAGLE_START
+# @param category Memory
+# @param tag daliuge
+# @param data_volume Data volume/5/Float/ComponentParameter/readwrite//False/False/Estimated size of the data contained in this node
+# @param group_end Group end/False/Boolean/ComponentParameter/readwrite//False/False/Is this node the end of a group?
+# @param dummy dummy//Object/InputPort/readwrite//False/False/Dummy input port
+# @param dummy dummy//Object/OutputPort/readwrite//False/False/Dummy output port
+# @par EAGLE_END
+class InMemoryDROP(DataDROP):
+    """
+    A DROP that points data stored in memory.
+    """
+
+    # Allow in-memory drops to be automatically removed by default
+    def __init__(self, *args, **kwargs):
+        if "precious" not in kwargs:
+            kwargs["precious"] = False
+        if "expireAfterUse" not in kwargs:
+            kwargs["expireAfterUse"] = True
+        super().__init__(*args, **kwargs)
+
+    def initialize(self, **kwargs):
+        args = []
+        if "pydata" in kwargs:
+            pydata = kwargs.pop("pydata")
+            if isinstance(pydata, str):
+                pydata = pydata.encode("utf8")
+            args.append(base64.b64decode(pydata))
+        self._buf = io.BytesIO(*args)
+
+    def getIO(self):
+        if (
+            hasattr(self, "_tp")
+            and hasattr(self, "_sessID")
+            and sys.version_info >= (3, 8)
+        ):
+            return SharedMemoryIO(self.oid, self._sessID)
+        else:
+            return MemoryIO(self._buf)
+
+    @property
+    def dataURL(self) -> str:
+        hostname = os.uname()[1]
+        return "mem://%s/%d/%d" % (hostname, os.getpid(), id(self._buf))
+
+    # Override
+    def generate_reproduce_data(self):
+        from dlg.droputils import allDropContents
+
+        data = b""
+        try:
+            data = allDropContents(self, self.size)
+        except Exception:
+            logger.debug("Could not read drop reproduce data")
+        return {"data_hash": common_hash(data)}
+
+
+##
+# @brief SharedMemory
+# @details Data stored in shared memory
+# @par EAGLE_START
+# @param category SharedMemory
+# @param tag template
+# @param data_volume Data volume/5/Float/ComponentParameter/readwrite//False/False/Estimated size of the data contained in this node
+# @param group_end Group end/False/Boolean/ComponentParameter/readwrite//False/False/Is this node the end of a group?
+# @param dummy dummy//Object/InputPort/readwrite//False/False/Dummy input port
+# @param dummy dummy//Object/OutputPort/readwrite//False/False/Dummy output port
+# @par EAGLE_END
+class SharedMemoryDROP(DataDROP):
+    """
+    A DROP that points to data stored in shared memory.
+    This drop is functionality equivalent to an InMemory drop running in a concurrent environment.
+    In this case however, the requirement for shared memory is explicit.
+
+    @WARNING Currently implemented as writing to shmem and there is no backup behaviour.
+    """
+
+    def initialize(self, **kwargs):
+        args = []
+        if "pydata" in kwargs:
+            pydata = kwargs.pop("pydata")
+            if isinstance(pydata, str):
+                pydata = pydata.encode("utf8")
+            args.append(base64.b64decode(pydata))
+        self._buf = io.BytesIO(*args)
+
+    def getIO(self):
+        if sys.version_info >= (3, 8):
+            if hasattr(self, "_sessID"):
+                return SharedMemoryIO(self.oid, self._sessID)
+            else:
+                # Using Drop without manager, just generate a random name.
+                sess_id = "".join(
+                    random.choices(string.ascii_uppercase + string.digits, k=10)
+                )
+                return SharedMemoryIO(self.oid, sess_id)
+        else:
+            raise NotImplementedError(
+                "Shared memory is only available with Python >= 3.8"
+            )
+
+    @property
+    def dataURL(self) -> str:
+        hostname = os.uname()[1]
+        return f"shmem://{hostname}/{os.getpid()}/{id(self._buf)}"
