+++ daliuge-engine/dlg/drop.py
@@ -23,30 +23,21 @@
 Module containing the core DROP classes.
 """
 from sqlite3 import OperationalError
-import string
 from abc import ABCMeta, abstractmethod, abstractproperty
 import ast
-import base64
 from collections import OrderedDict
-import contextlib
-import errno
 import heapq
-import importlib
 import inspect
-import io
 import logging
 import math
 import os
 import random
-import shutil
 import threading
 import time
 import re
 import sys
-import binascii
 from typing import List, Union
 
-import numpy as np
 from dlg.common.reproducibility.constants import (
     ReproducibilityFlags,
     REPRO_DEFAULT,
@@ -70,14 +61,8 @@ from dlg.exceptions import InvalidDropException, InvalidRelationshipException
 from dlg.io import (
     DataIO,
     OpenMode,
-    FileIO,
-    MemoryIO,
-    NgasIO,
-    NgasLiteIO,
     ErrorIO,
     NullIO,
-    PlasmaIO,
-    PlasmaFlightIO,
 )
 
 DEFAULT_INTERNAL_PARAMETERS = {
@@ -93,9 +78,8 @@ DEFAULT_INTERNAL_PARAMETERS = {
 }
 
 if sys.version_info >= (3, 8):
-    from dlg.io import SharedMemoryIO
+    pass
 from dlg.utils import (
-    prepare_sql,
     createDirIfMissing,
     isabs,
     object_tracking,
@@ -111,8 +95,6 @@ from dlg.meta import (
     dlg_bool_param,
     dlg_dict_param,
 )
-
-import pyarrow.plasma as plasma
 
 # Opt into using per-drop checksum calculation
 checksum_disabled = "DLG_DISABLE_CHECKSUM" in os.environ
@@ -1486,431 +1468,6 @@ class DataDROP(AbstractDROP):
 
 
 ##
-# @brief File
-# @details A standard file on a filesystem mounted to the deployment machine
-# @par EAGLE_START
-# @param category File
-# @param tag daliuge
-# @param data_volume Data volume/5/Float/ComponentParameter/readwrite//False/False/Estimated size of the data contained in this node
-# @param group_end Group end/False/Boolean/ComponentParameter/readwrite//False/False/Is this node the end of a group?
-# @param check_filepath_exists Check file path exists/True/Boolean/ComponentParameter/readwrite//False/False/Perform a check to make sure the file path exists before proceeding with the application
-# @param filepath File Path//String/ComponentParameter/readwrite//False/False/Path to the file for this node
-# @param dirname Directory name//String/ComponentParameter/readwrite//False/False/Path to the file for this node
-# @param dummy dummy//Object/InputPort/readwrite//False/False/Dummy input port
-# @param dummy dummy//Object/OutputPort/readwrite//False/False/Dummy output port
-# @par EAGLE_END
-class FileDROP(DataDROP, PathBasedDrop):
-    """
-    A DROP that points to data stored in a mounted filesystem.
-
-    Users can (but usually don't need to) specify both a `filepath` and a
-    `dirname` parameter for each FileDrop. The combination of these two parameters
-    will determine the final location of the file backed up by this drop on the
-    underlying filesystem. When no ``filepath`` is provided, the drop's UID will be
-    used as a filename. When a relative filepath is provided, it is relative to
-    ``dirname``. When an absolute ``filepath`` is given, it is used as-is.
-    When a relative ``dirname`` is provided, it is relative to the base directory
-    of the currently running session (i.e., a directory with the session ID as a
-    name, placed within the currently working directory of the Node Manager
-    hosting that session). If ``dirname`` is absolute, it is used as-is.
-
-    In some cases drops are created **outside** the context of a session, most
-    notably during unit tests. In these cases the base directory is a fixed
-    location under ``/tmp``.
-
-    The following table summarizes the calculation of the final path used by
-    the ``FileDrop`` class depending on its parameters:
-
-    ============ ===================== ===================== ==========
-         .                               filepath
-    ------------ ------------------------------------------------------
-    dirname      empty                 relative              absolute
-    ============ ===================== ===================== ==========
-    **empty**    /``$B``/``$u``        /``$B``/``$f``        /``$f``
-    **relative** /``$B``/``$d``/``$u`` /``$B``/``$d``/``$f`` **ERROR**
-    **absolute** /``$d``/``$u``        /``$d``/``$f``        **ERROR**
-    ============ ===================== ===================== ==========
-
-    In the table, ``$f`` is the value of ``filepath``, ``$d`` is the value of
-    ``dirname``, ``$u`` is the drop's UID and ``$B`` is the base directory for
-    this drop's session, namely ``/the/cwd/$session_id``.
-    """
-
-    # filepath = dlg_string_param("filepath", None)
-    # dirname = dlg_string_param("dirname", None)
-    delete_parent_directory = dlg_bool_param("delete_parent_directory", False)
-    check_filepath_exists = dlg_bool_param("check_filepath_exists", False)
-
-    def sanitize_paths(self, filepath, dirname):
-
-        # first replace any ENV_VARS on the names
-        if filepath:
-            filepath = os.path.expandvars(filepath)
-        if dirname:
-            dirname = os.path.expandvars(dirname)
-        # No filepath has been given, there's nothing to sanitize
-        if not filepath:
-            return filepath, dirname
-
-        # All is good, return unchanged
-        filepath_b = os.path.basename(filepath)
-        if filepath_b == filepath:
-            return filepath, dirname
-
-        # Extract the dirname from filepath and append it to dirname
-        filepath_d = os.path.dirname(filepath)
-        if not isabs(filepath_d) and dirname:
-            filepath_d = os.path.join(dirname, filepath_d)
-        return filepath_b, filepath_d
-
-    non_fname_chars = re.compile(r":|%s" % os.sep)
-
-    def initialize(self, **kwargs):
-        """
-        FileDROP-specific initialization.
-        """
-        # filepath, dirpath the two pieces of information we offer users to tweak
-        # These are very intermingled but are not exactly the same, see below
-        self.filepath = self.parameters.get("filepath", None)
-        self.dirname = self.parameters.get("dirname", None)
-        # Duh!
-        if isabs(self.filepath) and self.dirname:
-            raise InvalidDropException(
-                self, "An absolute filepath does not allow a dirname to be specified"
-            )
-
-        # Sanitize filepath/dirname into proper directories-only and
-        # filename-only components (e.g., dirname='lala' and filename='1/2'
-        # results in dirname='lala/1' and filename='2'
-        filepath, dirname = self.sanitize_paths(self.filepath, self.dirname)
-        # We later check if the file exists, but only if the user has specified
-        # an absolute dirname/filepath (otherwise it doesn't make sense, since
-        # we create our own filenames/dirnames dynamically as necessary
-        check = False
-        if isabs(dirname) and filepath:
-            check = self.check_filepath_exists
-
-        # Default filepath to drop UID and dirname to per-session directory
-        if not filepath:
-            filepath = self.non_fname_chars.sub("_", self.uid)
-        dirname = self.get_dir(dirname)
-
-        self._root = dirname
-        self._path = os.path.join(dirname, filepath)
-        logger.debug(f"Set path of drop {self._uid}: {self._path}")
-        if check and not os.path.isfile(self._path):
-            raise InvalidDropException(
-                self, "File does not exist or is not a file: %s" % self._path
-            )
-
-        self._wio = None
-
-    def getIO(self):
-        return FileIO(self._path)
-
-    def delete(self):
-        super().delete()
-        if self.delete_parent_directory:
-            try:
-                os.rmdir(self._root)
-            except OSError as e:
-                # Silently ignore "Directory not empty" errors
-                if e.errno != errno.ENOTEMPTY:
-                    raise
-
-    @track_current_drop
-    def setCompleted(self):
-        """
-        Override this method in order to get the size of the drop set once it is completed.
-        """
-        # TODO: This implementation is almost a verbatim copy of the base class'
-        # so we should look into merging them
-        status = self.status
-        if status == DROPStates.CANCELLED:
-            return
-        elif status == DROPStates.SKIPPED:
-            self._fire("dropCompleted", status=status)
-            return
-        elif status not in [DROPStates.INITIALIZED, DROPStates.WRITING]:
-            raise Exception(
-                "%r not in INITIALIZED or WRITING state (%s), cannot setComplete()"
-                % (self, self.status)
-            )
-
-        self._closeWriters()
-
-        logger.debug("Moving %r to COMPLETED", self)
-        self.status = DROPStates.COMPLETED
-
-        # here we set the size. It could happen that nothing is written into
-        # this file, in which case we create an empty file so applications
-        # downstream don't fail to read
-        try:
-            self._size = os.stat(self.path).st_size
-        except FileNotFoundError:
-            # we''ll try this again in case there is some other issue
-            try:
-                with open(self.path, "wb"):
-                    pass
-            except:
-                self.status = DROPStates.ERROR
-                logger.error("Path not accessible: %s", self.path)
-            self._size = 0
-        # Signal our subscribers that the show is over
-        self._fire("dropCompleted", status=DROPStates.COMPLETED)
-        self.completedrop()
-
-    @property
-    def dataURL(self) -> str:
-        hostname = os.uname()[1]  # TODO: change when necessary
-        return "file://" + hostname + self._path
-
-    # Override
-    def generate_reproduce_data(self):
-        from .droputils import allDropContents
-
-        try:
-            data = allDropContents(self, self.size)
-        except Exception:
-            data = b""
-        return {"data_hash": common_hash(data)}
-
-
-##
-# @brief NGAS
-# @details An archive on the Next Generation Archive System (NGAS).
-# @par EAGLE_START
-# @param category NGAS
-# @param tag daliuge
-# @param data_volume Data volume/5/Float/ComponentParameter/readwrite//False/False/Estimated size of the data contained in this node
-# @param group_end Group end/False/Boolean/ComponentParameter/readwrite//False/False/Is this node the end of a group?
-# @param ngasSrv NGAS Server/localhost/String/ComponentParameter/readwrite//False/False/The URL of the NGAS Server
-# @param ngasPort NGAS Port/7777/Integer/ComponentParameter/readwrite//False/False/The port of the NGAS Server
-# @param ngasFileId File ID//String/ComponentParameter/readwrite//False/False/File ID on NGAS (for retrieval only)
-# @param ngasConnectTimeout Connection timeout/2/Integer/ComponentParameter/readwrite//False/False/Timeout for connecting to the NGAS server
-# @param ngasMime NGAS mime-type/"text/ascii"/String/ComponentParameter/readwrite//False/False/Mime-type to be used for archiving
-# @param ngasTimeout NGAS timeout/2/Integer/ComponentParameter/readwrite//False/False/Timeout for receiving responses for NGAS
-# @param dummy dummy//Object/InputPort/readwrite//False/False/Dummy input port
-# @param dummy dummy//Object/OutputPort/readwrite//False/False/Dummy output port
-# @par EAGLE_END
-class NgasDROP(DataDROP):
-    """
-    A DROP that points to data stored in an NGAS server
-    """
-
-    ngasSrv = dlg_string_param("ngasSrv", "localhost")
-    ngasPort = dlg_int_param("ngasPort", 7777)
-    ngasFileId = dlg_string_param("ngasFileId", None)
-    ngasTimeout = dlg_int_param("ngasTimeout", 2)
-    ngasConnectTimeout = dlg_int_param("ngasConnectTimeout", 2)
-    ngasMime = dlg_string_param("ngasMime", "application/octet-stream")
-    len = dlg_int_param("len", -1)
-    ngas_checksum = None
-
-    def initialize(self, **kwargs):
-        if self.len == -1:
-            # TODO: For writing the len field should be set to the size of the input drop
-            self.len = self._size
-        if self.ngasFileId:
-            self.fileId = self.ngasFileId
-        else:
-            self.fileId = self.uid
-
-    def getIO(self):
-        try:
-            ngasIO = NgasIO(
-                self.ngasSrv,
-                self.fileId,
-                self.ngasPort,
-                self.ngasConnectTimeout,
-                self.ngasTimeout,
-                length=self.len,
-                mimeType=self.ngasMime,
-            )
-        except ImportError:
-            logger.warning("NgasIO not available, using NgasLiteIO instead")
-            ngasIO = NgasLiteIO(
-                self.ngasSrv,
-                self.fileId,
-                self.ngasPort,
-                self.ngasConnectTimeout,
-                self.ngasTimeout,
-                length=self.len,
-                mimeType=self.ngasMime,
-            )
-        return ngasIO
-
-    @track_current_drop
-    def setCompleted(self):
-        """
-        Override this method in order to get the size of the drop set once it is completed.
-        """
-        # TODO: This implementation is almost a verbatim copy of the base class'
-        # so we should look into merging them
-        status = self.status
-        if status == DROPStates.CANCELLED:
-            return
-        elif status == DROPStates.SKIPPED:
-            self._fire("dropCompleted", status=status)
-            return
-        elif status not in [DROPStates.INITIALIZED, DROPStates.WRITING]:
-            raise Exception(
-                "%r not in INITIALIZED or WRITING state (%s), cannot setComplete()"
-                % (self, self.status)
-            )
-
-        self._closeWriters()
-
-        # here we set the size. It could happen that nothing is written into
-        # this file, in which case we create an empty file so applications
-        # downstream don't fail to read
-        logger.debug("Trying to set size of NGASDrop")
-        try:
-            stat = self.getIO().fileStatus()
-            logger.debug(
-                "Setting size of NGASDrop %s to %s", self.fileId, stat["FileSize"]
-            )
-            self._size = int(stat["FileSize"])
-            self.ngas_checksum = str(stat["Checksum"])
-        except:
-            # we''ll try this again in case there is some other issue
-            # try:
-            #     with open(self.path, 'wb'):
-            #         pass
-            # except:
-            #     self.status = DROPStates.ERROR
-            #     logger.error("Path not accessible: %s" % self.path)
-            raise
-            logger.debug("Setting size of NGASDrop to %s", 0)
-            self._size = 0
-        # Signal our subscribers that the show is over
-        logger.debug("Moving %r to COMPLETED", self)
-        self.status = DROPStates.COMPLETED
-        self._fire("dropCompleted", status=DROPStates.COMPLETED)
-        self.completedrop()
-
-    @property
-    def dataURL(self) -> str:
-        return "ngas://%s:%d/%s" % (self.ngasSrv, self.ngasPort, self.fileId)
-
-    # Override
-    def generate_reproduce_data(self):
-        if self.ngas_checksum is None or self.ngas_checksum == '':
-            return {"fileid": self.ngasFileId, "size": self._size}
-        return {"data_hash": self.ngas_checksum}
-
-
-##
-# @brief Memory
-# @details In-memory storage of intermediate data products
-# @par EAGLE_START
-# @param category Memory
-# @param tag daliuge
-# @param data_volume Data volume/5/Float/ComponentParameter/readwrite//False/False/Estimated size of the data contained in this node
-# @param group_end Group end/False/Boolean/ComponentParameter/readwrite//False/False/Is this node the end of a group?
-# @param dummy dummy//Object/InputPort/readwrite//False/False/Dummy input port
-# @param dummy dummy//Object/OutputPort/readwrite//False/False/Dummy output port
-# @par EAGLE_END
-class InMemoryDROP(DataDROP):
-    """
-    A DROP that points data stored in memory.
-    """
-
-    # Allow in-memory drops to be automatically removed by default
-    def __init__(self, *args, **kwargs):
-        if 'precious' not in kwargs:
-            kwargs['precious'] = False
-        if 'expireAfterUse' not in kwargs:
-            kwargs['expireAfterUse'] = True
-        super().__init__(*args, **kwargs)
-
-    def initialize(self, **kwargs):
-        args = []
-        if "pydata" in kwargs:
-            pydata = kwargs.pop("pydata")
-            if isinstance(pydata, str):
-                pydata = pydata.encode("utf8")
-            args.append(base64.b64decode(pydata))
-        self._buf = io.BytesIO(*args)
-
-    def getIO(self):
-        if (
-            hasattr(self, "_tp")
-            and hasattr(self, "_sessID")
-            and sys.version_info >= (3, 8)
-        ):
-            return SharedMemoryIO(self.oid, self._sessID)
-        else:
-            return MemoryIO(self._buf)
-
-    @property
-    def dataURL(self) -> str:
-        hostname = os.uname()[1]
-        return "mem://%s/%d/%d" % (hostname, os.getpid(), id(self._buf))
-
-    # Override
-    def generate_reproduce_data(self):
-        from .droputils import allDropContents
-
-        data = b""
-        try:
-            data = allDropContents(self, self.size)
-        except Exception:
-            logger.debug("Could not read drop reproduce data")
-        return {"data_hash": common_hash(data)}
-
-
-##
-# @brief SharedMemory
-# @details Data stored in shared memory
-# @par EAGLE_START
-# @param category SharedMemory
-# @param tag template
-# @param data_volume Data volume/5/Float/ComponentParameter/readwrite//False/False/Estimated size of the data contained in this node
-# @param group_end Group end/False/Boolean/ComponentParameter/readwrite//False/False/Is this node the end of a group?
-# @param dummy dummy//Object/InputPort/readwrite//False/False/Dummy input port
-# @param dummy dummy//Object/OutputPort/readwrite//False/False/Dummy output port
-# @par EAGLE_END
-class SharedMemoryDROP(DataDROP):
-    """
-    A DROP that points to data stored in shared memory.
-    This drop is functionality equivalent to an InMemory drop running in a concurrent environment.
-    In this case however, the requirement for shared memory is explicit.
-
-    @WARNING Currently implemented as writing to shmem and there is no backup behaviour.
-    """
-
-    def initialize(self, **kwargs):
-        args = []
-        if "pydata" in kwargs:
-            pydata = kwargs.pop("pydata")
-            if isinstance(pydata, str):
-                pydata = pydata.encode("utf8")
-            args.append(base64.b64decode(pydata))
-        self._buf = io.BytesIO(*args)
-
-    def getIO(self):
-        if sys.version_info >= (3, 8):
-            if hasattr(self, "_sessID"):
-                return SharedMemoryIO(self.oid, self._sessID)
-            else:
-                # Using Drop without manager, just generate a random name.
-                sess_id = "".join(
-                    random.choices(string.ascii_uppercase + string.digits, k=10)
-                )
-                return SharedMemoryIO(self.oid, sess_id)
-        else:
-            raise NotImplementedError(
-                "Shared memory is only available with Python >= 3.8"
-            )
-
-    @property
-    def dataURL(self) -> str:
-        hostname = os.uname()[1]
-        return f"shmem://{hostname}/{os.getpid()}/{id(self._buf)}"
-
-
-##
 # @brief NULL
 # @details A Drop not storing any data (useful for just passing on events)
 # @par EAGLE_START
@@ -1938,122 +1495,6 @@ class EndDROP(NullDROP):
     """
     A DROP that ends the session when reached
     """
-
-
-##
-# @brief RDBMS
-# @details A Drop allowing storage and retrieval from a SQL DB.
-# @par EAGLE_START
-# @param category File
-# @param tag template
-# @param data_volume Data volume/5/Float/ComponentParameter/readwrite//False/False/Estimated size of the data contained in this node
-# @param group_end Group end/False/Boolean/ComponentParameter/readwrite//False/False/Is this node the end of a group?
-# @param dbmodule Python DB module//String/ComponentParameter/readwrite//False/False/Load path for python DB module
-# @param dbtable DB table name//String/ComponentParameter/readwrite//False/False/The name of the table to use
-# @param vals Values dictionary/{}/Json/ComponentParameter/readwrite//False/False/Json encoded values dictionary used for INSERT. The keys of ``vals`` are used as the column names.
-# @param condition Whats used after WHERE//String/ComponentParameter/readwrite//False/False/Condition for SELECT. For this the WHERE statement must be written using the "{X}" or "{}" placeholders
-# @param selectVals values for WHERE/{}/Json/ComponentParameter/readwrite//False/False/Values for the WHERE statement
-# @param dummy dummy//Object/InputPort/readwrite//False/False/Dummy input port
-# @param dummy dummy//Object/OutputPort/readwrite//False/False/Dummy output port
-# @par EAGLE_END
-class RDBMSDrop(DataDROP):
-    """
-    A Drop that stores data in a table of a relational database
-    """
-
-    dbparams = dlg_dict_param("dbparams", {})
-
-    def initialize(self, **kwargs):
-        DataDROP.initialize(self, **kwargs)
-
-        if "dbmodule" not in kwargs:
-            raise InvalidDropException(
-                self, '%r needs a "dbmodule" parameter' % (self,)
-            )
-        if "dbtable" not in kwargs:
-            raise InvalidDropException(self, '%r needs a "dbtable" parameter' % (self,))
-
-        # The DB-API 2.0 module
-        dbmodname = kwargs.pop("dbmodule")
-        self._db_drv = importlib.import_module(dbmodname)
-
-        # The table this Drop points at
-        self._db_table = kwargs.pop("dbtable")
-
-        # Data store for reproducibility
-        self._querylog = []
-
-    def getIO(self):
-        # This Drop cannot be accessed directly
-        return ErrorIO()
-
-    def _connection(self):
-        return contextlib.closing(self._db_drv.connect(**self.dbparams))
-
-    def _cursor(self, conn):
-        return contextlib.closing(conn.cursor())
-
-    def insert(self, vals: dict):
-        """
-        Inserts the values contained in the ``vals`` dictionary into the
-        underlying table. The keys of ``vals`` are used as the column names.
-        """
-        with self._connection() as c:
-            with self._cursor(c) as cur:
-                # vals is a dictionary, its keys are the column names and its
-                # values are the values to insert
-                sql = "INSERT into %s (%s) VALUES (%s)" % (
-                    self._db_table,
-                    ",".join(vals.keys()),
-                    ",".join(["{}"] * len(vals)),
-                )
-                sql, vals = prepare_sql(
-                    sql, self._db_drv.paramstyle, list(vals.values())
-                )
-                logger.debug("Executing SQL with parameters: %s / %r", sql, vals)
-                cur.execute(sql, vals)
-                c.commit()
-
-    def select(self, columns=None, condition=None, vals=()):
-        """
-        Returns the selected values from the table. Users can constrain the
-        result set by specifying a list of ``columns`` to be returned (otherwise
-        all table columns are returned) and a ``condition`` to be applied,
-        in which case a list of ``vals`` to be applied as query parameters can
-        also be given.
-        """
-        with self._connection() as c:
-            with self._cursor(c) as cur:
-
-                # Build up SQL with optional columns and conditions
-                columns = columns or ("*",)
-                sql = ["SELECT %s FROM %s" % (",".join(columns), self._db_table)]
-                if condition:
-                    sql.append(" WHERE ")
-                    sql.append(condition)
-
-                # Go, go, go!
-                sql, vals = prepare_sql("".join(sql), self._db_drv.paramstyle, vals)
-                logger.debug("Executing SQL with parameters: %s / %r", sql, vals)
-                cur.execute(sql, vals)
-                if cur.description:
-                    ret = cur.fetchall()
-                else:
-                    ret = []
-                self._querylog.append((sql, vals, ret))
-                return ret
-
-    @property
-    def dataURL(self) -> str:
-        return "rdbms://%s/%s/%r" % (
-            self._db_drv.__name__,
-            self._db_table,
-            self._db_params,
-        )
-
-    # Override
-    def generate_reproduce_data(self):
-        return {"query_log": self._querylog}
 
 
 class ContainerDROP(DataDROP):
@@ -2119,147 +1560,6 @@ class ContainerDROP(DataDROP):
             #       "exists" is
             return any([c.exists() for c in self._children])
         return True
-
-
-class DirectoryContainer(PathBasedDrop, ContainerDROP):
-    """
-    A ContainerDROP that represents a filesystem directory. It only allows
-    FileDROPs and DirectoryContainers to be added as children. Children
-    can only be added if they are placed directly within the directory
-    represented by this DirectoryContainer.
-    """
-
-    check_exists = dlg_bool_param("check_exists", True)
-
-    def initialize(self, **kwargs):
-        ContainerDROP.initialize(self, **kwargs)
-
-        if "dirname" not in kwargs:
-            raise InvalidDropException(
-                self, 'DirectoryContainer needs a "dirname" parameter'
-            )
-
-        directory = kwargs["dirname"]
-
-        if self.check_exists is True:
-            if not os.path.isdir(directory):
-                raise InvalidDropException(self, "%s is not a directory" % (directory))
-
-        self._path = self.get_dir(directory)
-
-    def addChild(self, child):
-        if isinstance(child, (FileDROP, DirectoryContainer)):
-            path = child.path
-            if os.path.dirname(path) != self.path:
-                raise InvalidRelationshipException(
-                    DROPRel(child, DROPLinkType.CHILD, self),
-                    "Child DROP is not under %s" % (self.path),
-                )
-            ContainerDROP.addChild(self, child)
-        else:
-            raise TypeError("Child DROP is not of type FileDROP or DirectoryContainer")
-
-    def delete(self):
-        shutil.rmtree(self._path)
-
-    def exists(self):
-        return os.path.isdir(self._path)
-
-
-##
-# @brief Plasma
-# @details An object in a Apache Arrow Plasma in-memory object store
-# @par EAGLE_START
-# @param category Plasma
-# @param tag daliuge
-# @param data_volume Data volume/5/Float/ComponentParameter/readwrite//False/False/Estimated size of the data contained in this node
-# @param group_end Group end/False/Boolean/ComponentParameter/readwrite//False/False/Is this node the end of a group?
-# @param plasma_path Plasma Path//String/ComponentParameter/readwrite//False/False/Path to the local plasma store
-# @param object_id Object Id//String/ComponentParameter/readwrite//False/False/PlasmaId of the object for all compute nodes
-# @param use_staging Use Staging/False/Boolean/ComponentParameter/readwrite//False/False/Enables writing to a dynamically resizeable staging buffer
-# @param dummy dummy//Object/InputPort/readwrite//False/False/Dummy input port
-# @param dummy dummy//Object/OutputPort/readwrite//False/False/Dummy output port
-# @par EAGLE_END
-class PlasmaDROP(DataDROP):
-    """
-    A DROP that points to data stored in a Plasma Store
-    """
-
-    object_id: bytes = dlg_string_param("object_id", None)
-    plasma_path: str = dlg_string_param("plasma_path", "/tmp/plasma")
-    use_staging: bool = dlg_bool_param("use_staging", False)
-
-    def initialize(self, **kwargs):
-        super().initialize(**kwargs)
-        self.plasma_path = os.path.expandvars(self.plasma_path)
-        if self.object_id is None:
-            self.object_id = (
-                np.random.bytes(20) if len(self.uid) != 20 else self.uid.encode("ascii")
-            )
-        elif isinstance(self.object_id, str):
-            self.object_id = self.object_id.encode("ascii")
-
-    def getIO(self):
-        return PlasmaIO(
-            plasma.ObjectID(self.object_id),
-            self.plasma_path,
-            expected_size=self._expectedSize,
-            use_staging=self.use_staging,
-        )
-
-    @property
-    def dataURL(self) -> str:
-        return "plasma://%s" % (binascii.hexlify(self.object_id).decode("ascii"))
-
-
-##
-# @brief PlasmaFlight
-# @details An Apache Arrow Flight server providing distributed access
-# to a Plasma in-memory object store
-# @par EAGLE_START
-# @param category PlasmaFlight
-# @param tag daliuge
-# @param data_volume Data volume/5/Float/ComponentParameter/readwrite//False/False/Estimated size of the data contained in this node
-# @param group_end Group end/False/Boolean/ComponentParameter/readwrite//False/False/Is this node the end of a group?
-# @param plasma_path Plasma Path//String/ComponentParameter/readwrite//False/False/Path to the local plasma store
-# @param object_id Object Id//String/ComponentParameter/readwrite//False/False/PlasmaId of the object for all compute nodes
-# @param flight_path Flight Path//String/ComponentParameter/readwrite//False/False/IP and flight port of the drop owner
-# @param dummy dummy//Object/InputPort/readwrite//False/False/Dummy input port
-# @param dummy dummy//Object/OutputPort/readwrite//False/False/Dummy output port
-# @par EAGLE_END
-class PlasmaFlightDROP(DataDROP):
-    """
-    A DROP that points to data stored in a Plasma Store
-    """
-
-    object_id: bytes = dlg_string_param("object_id", None)
-    plasma_path: str = dlg_string_param("plasma_path", "/tmp/plasma")
-    flight_path: str = dlg_string_param("flight_path", None)
-    use_staging: bool = dlg_bool_param("use_staging", False)
-
-    def initialize(self, **kwargs):
-        super().initialize(**kwargs)
-        self.plasma_path = os.path.expandvars(self.plasma_path)
-        if self.object_id is None:
-            self.object_id = (
-                np.random.bytes(20) if len(self.uid) != 20 else self.uid.encode("ascii")
-            )
-        elif isinstance(self.object_id, str):
-            self.object_id = self.object_id.encode("ascii")
-
-    def getIO(self):
-        return PlasmaFlightIO(
-            plasma.ObjectID(self.object_id),
-            self.plasma_path,
-            flight_path=self.flight_path,
-            expected_size=self._expectedSize,
-            use_staging=self.use_staging,
-        )
-
-    @property
-    def dataURL(self) -> str:
-        return "plasmaflight://%s" % (binascii.hexlify(self.object_id).decode("ascii"))
-
 
 # ===============================================================================
 # AppDROP classes follow
@@ -2684,41 +1984,6 @@ class BarrierAppDROP(InputFiredAppDROP):
         # Blindly override existing value if any
         kwargs["n_effective_inputs"] = -1
         super().initialize(**kwargs)
-
-
-##
-# @brief Branch
-# @details A conditional branch to control flow
-# @par EAGLE_START
-# @param category Branch
-# @param tag template
-# @param appclass Application Class/dlg.apps.simple.SimpleBranch/String/ComponentParameter/readonly//False/False/Application class
-# @param execution_time Execution Time/5/Float/ComponentParameter/readonly//False/False/Estimated execution time
-# @param num_cpus No. of CPUs/1/Integer/ComponentParameter/readonly//False/False/Number of cores used
-# @param group_start Group start/False/Boolean/ComponentParameter/readwrite//False/False/Is this node the start of a group?
-# @param input_error_threshold "Input error rate (%)"/0/Integer/ComponentParameter/readwrite//False/False/the allowed failure rate of the inputs (in percent), before this component goes to ERROR state and is not executed
-# @param n_tries Number of tries/1/Integer/ComponentParameter/readwrite//False/False/Specifies the number of times the 'run' method will be executed before finally giving up
-# @param dummy0 dummy0//Object/OutputPort/readwrite//False/False/Dummy output port
-# @param dummy1 dummy1//Object/OutputPort/readwrite//False/False/Dummy output port
-# @par EAGLE_END
-class BranchAppDrop(BarrierAppDROP):
-    """
-    A special kind of application with exactly two outputs. After normal
-    execution, the application decides whether a certain condition is met.
-    If the condition is met, the first output is considered as COMPLETED,
-    while the other is moved to SKIPPED state, and vice-versa.
-    """
-
-    @track_current_drop
-    def execute(self, _send_notifications=True):
-        if len(self._outputs) != 2:
-            raise InvalidDropException(
-                self,
-                f"BranchAppDrops should have exactly 2 outputs, not {len(self._outputs)}",
-            )
-        BarrierAppDROP.execute(self, _send_notifications=False)
-        self.outputs[1 if self.condition() else 0].skip()
-        self._notifyAppIsFinished()
 
 
 # Dictionary mapping 1-to-many DROPLinkType constants to the corresponding methods
