+++ daliuge-engine/test/apps/test_bash.py
@@ -36,7 +36,8 @@ from dlg.apps.bash_shell_app import (
     StreamingInputOutputBashApp,
 )
 from dlg.ddap_protocol import DROPStates
-from dlg.drop import FileDROP, InMemoryDROP
+from dlg.data.memory import InMemoryDROP
+from dlg.data.file import FileDROP
 from dlg.droputils import DROPWaiterCtx
