+++ daliuge-engine/test/apps/test_docker.py
@@ -31,7 +31,8 @@ import docker
 
 from dlg import droputils, utils
 from dlg.apps.dockerapp import DockerApp
-from dlg.drop import FileDROP, NgasDROP
+from dlg.data.ngas import NgasDROP
+from dlg.data.file import FileDROP
 from dlg.droputils import DROPWaiterCtx
 
 docker_unavailable = True
