+++ daliuge-engine/test/lifecycle/test_registry.py
@@ -23,7 +23,7 @@ import os
 import sqlite3
 import unittest
 
-from dlg.drop import InMemoryDROP
+from dlg.data.memory import InMemoryDROP
 from dlg.lifecycle.registry import RDBMSRegistry
